requirements.txt
flask==3.0.3
tensorflow-cpu==2.15.0
pandas
scikit-learn
joblib
numpy
